added compare to sqare logic

+++ sfml_choke/objects/Game.hpp
@@ -22,10 +22,10 @@ namespace GO {
   {
   private:
     struct mapBlockData {
-      int width;
-      int height;
-      int columnsCount;
-      int linesCount;
+      double width;
+      double height;
+      double columnsCount;
+      double linesCount;
     };
     
     sf::RenderWindow* gameWindow;
